+++ src/CMS_default.java
@@ -11,13 +11,13 @@ public class CMS_default implements CMS {
 	private int width;
 	private int depth;
 	private Hash h[];
-	private int c[][];
+	private long c[][];
 
 	CMS_default(int d, int w) {
 		width = w;
 		depth = d;
 		h = new Hash[depth];
-		c = new int[depth][width];
+		c = new long[depth][width];
 		
 		for (int j =0; j<depth; ++j) {
 			h[j] = new Hash();
@@ -30,9 +30,9 @@ public class CMS_default implements CMS {
 		}
 	}
 
-	public int query(Object o) {
-		int min = c[0][h[0].h_basic(o, width)];
-		int current;
+	public long query(Object o) {
+		long min = c[0][h[0].h_basic(o, width)];
+		long current;
 		
 		for (int j=1; j < depth ; j++) {
 			current = c[j][h[j].h_basic(o, width)];
